+++ examples/mt4/src/main/java/MetaApiSynchronizationExample.java
@@ -42,7 +42,7 @@ public class MetaApiSynchronizationExample {
                     name = serverName;
                     version = 4;
                     brokerTimezone = "EET";
-                    brokerDSTTimezone = "EET";
+                    brokerDSTSwitchTimezone = "EET";
                 }};
                 profile = Optional.of(api.getProvisioningProfileApi().createProvisioningProfile(newDto).get());
                 profile.get().uploadFile("broker.srv", brokerSrvFile).get();