+++ changelog.md
@@ -5,9 +5,15 @@
   - made it possible to use MetaApi class in interaction tests
   - renamed tradeCopyingSlippageInPercentPoints -> tradeCopyingSlippageInBasisPoints in CopyFactory history API
   - added application setting to MetaApi class to make it possible to launch several MetaApi applications in parallel on the same account
-  - breaking change: removed the `timeConverter` field from the account, replaced it with `brokerTimezone` and `brokerDSTTimezone` fields in the provisioning profile instead
+  - breaking change: removed the `timeConverter` field from the account, replaced it with `brokerTimezone` and `brokerDSTSwitchTimezone` fields in the provisioning profile instead
   - added originalComment and clientId fields to MetatraderPosition
   - examples extracted into own maven projects
+  - fixed occasional fake synchronization timeouts in waitSynchronized method
+  - breaking change: changed API contract of MetaApiConnection.waitSynchronized method
+  - added tags for MetaApi accounts
+  - minor adjustments to equity calculation algorithm
+  - added the ability to set the start time for synchronization, used for tests
+  - resynchronize on lost synchronization packet to ensure local terminal state consistency
 
 7.2.0
   - Added time fields in broker timezone to objects